try to guess castling availability if it's not in FEN
Sometime the Forsyth-Edward notation may not contain any
information about whether castling is still possible. In that case
the position of kings and rooks are used to determine whether
castling is still possible. It might be inaccurate in some cases,
but it is better than the previous approach of always enabling it.