fix incomplete validation of Portable Game Notation
In some cases, if the PGN was missing some information, there
where two or more possibilities which piece could move to the
destination field. To avoid this, the check will now reject any
such situations where two or more pieces of the same type could
move to the destination field.