implement new variant of the mobility evaluator
The MobilityEvaluator class is now called LinearMobilityEvaluator.
The new RootMobilityEvaluator calculates the score for mobility,
too, but the score progresses slower than the score of the "old"
MobilityEvaluator. To put the scores into formulas:

LinearMobilityEvaluator:
  S = c * nom(p)

RootMobilityEvaluator:
  S = c * sqrt(nom(p))

where c is a constant factor greater than zero, and nom(p) is the
number of possible moves for the position p on the chess board.
The RootMobilityEvaluator basically provides an alternative for
mobility scores where mobility should not be the dominant score
when there are a lot of legal moves possible.