use constexpr string views for constant evaluator ids
Technically, there was nothing wrong with using std::strings, but
the speedbench binary segfaulted when built on GCC 15.1.0. (It
did run fine when built with GCC 14.2.0.) Runs with GDB indicated
it may have to do with static initialization going wrong
somewhere and passing null pointer to string creation for some
reason. Therefore, the switch to string views is done to avoid
that.