ci: replace windows-2019 by windows-2025 in GHA
The Windows Server 2019 images are deprecated, starting on
2025-06-01, and will be unsupported by 2025-06-30. See
<https://github.com/actions/runner-images/issues/12045>
for more information. Therefore, the `windows-2019` runner is
removed from the workflow and the `windows-2025` runner will be
used instead.

The downside is that both of those runners use the same version
of Visual Studio 2022, whereas the `windows-2019` runner had
Visual Studio 2019.